Move C code to separate file
